Add shared components for demo/production code reuse
- Create formatters.ts with shared date/time/address utilities
- Create RecordDetailView shared component for record detail pages
- Create ClientDetailView shared component for client detail pages
- Refactor demo and production detail pages to use shared components
- Reduce code duplication by ~1300 lines
