+++ web/src/components/clients/index.ts
@@ -1,0 +1,2 @@
+export { ClientDetailView } from './ClientDetailView';
+export type { ClientDetailViewProps, ClientForDetail } from './ClientDetailView';
