+++ web/src/components/records/index.ts
@@ -1,0 +1,2 @@
+export { RecordDetailView } from './RecordDetailView';
+export type { RecordDetailViewProps, VisitRecordForDetail } from './RecordDetailView';
